refactor : move fetchSeaPostidByCoordinates function (locationServices.js -> weatherService.js)
