+++ frontend/src/services/locationService.js
@@ -1,34 +1,3 @@
-import axios from "@/axios";
-
-const baseUrl = process.env.VUE_APP_BASE_URL;
-const apisealocBaseUrl = `${baseUrl}/api/get-seaweather`;
-
-export async function fetchSeaPostidByCoordinates(lat, lon) {
-  try {
-    const response = await axios.post(
-      apisealocBaseUrl,
-      new URLSearchParams({ lat: lat, lon: lon }).toString(), 
-      {
-        headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-        }
-      }
-    );
-
-    if (response.status === 200) {
-      return response.data.data;
-
-    } else {
-      return null
-    }
-    
-      
-  } catch (error) {
-      console.error("Error fetching sea weather data:", error);
-    return { error: error.message };
-  }
-}
-
 export function getCurrentLocation() {
     return new Promise((resolve, reject) => {
       if (!navigator.geolocation) {
